+++ packages/schemas/src/admin/index.ts
@@ -69,18 +69,3 @@ export const UPDATE_USER_SCHEMA = z.object({
   steamId: z.string().max(255),
   discordId: z.string().max(255),
 });
-
-export const IMPORT_CITIZENS_SCHEMA = z.object({
-  name: z.string().min(1).max(255),
-  surname: z.string().min(1).max(255),
-  gender: z.string().min(1).max(255),
-  ethnicity: z.string().min(1).max(255),
-  dateOfBirth: z.date().or(z.string().min(2)),
-  address: z.string().max(255).nullable().optional(),
-  eyeColor: z.string().max(255).nullable().optional(),
-  hairColor: z.string().max(255).nullable().optional(),
-  height: z.string().max(255).nullable().optional(),
-  weight: z.string().max(255).nullable().optional(),
-});
-
-export const IMPORT_CITIZENS_ARR = z.array(IMPORT_CITIZENS_SCHEMA).min(1);
